+++ data/blip.txt
@@ -1,2 +1,2 @@
 T | 0 | eat
-D | 0 | eat | 2023-05-21 14:00
+D | 0 | eat | 2023-05-06 10:23
