+++ src/main/java/blip/commands/Command.java
@@ -1,3 +1,9 @@
+package blip.commands;
+
+import blip.ui.*;
+import blip.tasks.*;
+import blip.storage.*;
+
 public abstract class Command {
     public abstract void execute(TaskList taskList, BlipUI ui, BlipStorage storage);
