+++ src/main/java/blip/commands/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package blip.commands;
+
+import blip.ui.*;
+import blip.tasks.*;
+import blip.storage.*;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
