+++ src/main/java/blip/commands/ListCommand.java
@@ -1,3 +1,9 @@
+package blip.commands;
+
+import blip.ui.*;
+import blip.tasks.*;
+import blip.storage.*;
+
 public class ListCommand extends Command {
     @Override
     public void execute(TaskList taskList, BlipUI ui, BlipStorage storage) {
