+++ src/main/java/blip/exceptions/BlipException.java
@@ -1,3 +1,4 @@
+package blip.exceptions;
 public class BlipException extends Exception {
     public BlipException(String message) { super(message); }
 }
