+++ src/main/java/blip/exceptions/EmptyDescriptionException.java
@@ -1,3 +1,4 @@
+package blip.exceptions;
 public class EmptyDescriptionException extends Exception {
 
     public EmptyDescriptionException(String message) { super(message); }
