+++ src/main/java/blip/exceptions/EmptyTaskNumberException.java
@@ -1,3 +1,4 @@
+package blip.exceptions;
 public class EmptyTaskNumberException extends Exception {
 
     public EmptyTaskNumberException(String message) {
