+++ src/main/java/blip/exceptions/InvalidCommandException.java
@@ -1,3 +1,4 @@
+package blip.exceptions;
 public class InvalidCommandException extends Exception {
 
     public InvalidCommandException(String message) { super(message); }
