+++ src/main/java/blip/exceptions/WrongNumberException.java
@@ -1,3 +1,4 @@
+package blip.exceptions;
 public class WrongNumberException extends Exception{
     public WrongNumberException(String message) {
         super(message);
