+++ src/main/java/blip/tasks/ToDo.java
@@ -1,3 +1,5 @@
+package blip.tasks;
+
 public class ToDo extends Task{
 
     public ToDo(String description, boolean isDone) {
